+++ mock/index.ts
@@ -1,5 +1,6 @@
 import auth from './auth'
 import user from './user'
 import dashboard from './dashboard'
+import menu from './menu'
 
-export default [...auth, ...user, ...dashboard]
+export default [...auth, ...user, ...dashboard, ...menu]